navigated to the analytics screen from the category card component
